Get auto-buy working
buy_value wasn't calculated correctly (wasn't dividing by 100) and
I got a quote precision error so I needed to round_de(sym.quote_precision).
But the quote_precision needs to be a u32 not a u64. So I had to add
the de_string_or_number_to_i32 and u32 and change the various _precision
value in struct Symbol. Fixed the output.

Also, added checking for asset != quote_asset so we ignore "USDUSD" and
the like. Not sure why I didn't run into that before. Added rounding
sell_qty to quote_precision in auto-sell as I did with buy_value. And
some cleanup of the output.
quote_precision in the