Added binance_history
Which displays the deposit and withdraw history of the account based
off binance_my_trades. I started down the path were I thought
HistoryRec was going to be the same for deposit and withdraw so
used one struct. But they are only "mostly" the same so this needs
to be refactored.

Added "dh", "wh" and "mt" to be displayed at towards the top of the help
and "ao" above "ol".

Modified mt subcommand to use the histories to calculate the
equivalent of Total Coins in from the ai subcommand. On my account
its pretty close for ETH. The ai subcommand reports 7.43502526 and
mt subcommand reports a little more, 7.43688064. This is a difference
of 0.0018554. My current guess is there are some transaction fees that
are not being accounted for.