Update to Edition 2021 and bump version to 0.3.0
Also update to the latest versions of the dependencies. I used
`cargo update`, `cargo upgrades`, `cargo upgrade` and `cargo outdated`.
Outdated did the "best" but it didn't update [dev-dependencies] and
[build-dependencies] to the very latest so I manually updated them.

The other goal of the updating to the latest versions was to eliminate
the duplicate versions of the same crate. Using `cargo tree -d` shows
what it thinks are duplicates but all of them are actually the same
version so I'm not sure what it's showing.

Add small tweaks to binance_signature and have it use new_from_slice
as new_varkey has been removed.

Finally, edition 2021 doesn't work on the old nightly and rustfmt is
now working on nightly. So toolchain is changed back to "nightly"
instead of "nightly-2021-03-25". Also, clippy complained about
additional items so fixed those in binance_auto_buy.rs,
binance_auto_sell.rs, common.rs and configuration.rs.