Add --keep-min to withdraw sub-command
This allows the user to specify a minimum amount of the asset to
NOT withdraw and instead keep in binacne.us.

Also allow the AMOUNT and --keep-min values be in percent, dollars
or quantiity. And pring the value withdrawn in USD.

NOTE: When dollars are used by prefixing with Dollar Sign, $. The value
must must escaped in single quotes or being preceeded with a backslash:

  '$456' or \$456

See binance-cli withdraw -h for more examples.