Change pdf to accept two parameters
The first parameter is the input file and is required, the second is
optional and is the output file. In future changes the second file will
be the same as the input file but also has fields that may have been
updated. Such as the missing USD amounts in some Distribution records.