Initial processing of every transaction
The code now porocess all transactions in all cateogories and operations
that are in binance.us-distribution-2021.csv.

By processing I mean each transaction is used and any missing USD fields
are updated using the closing value of the nearest minute kline.
There is a number of asserts in the code which validates that expected
counts are appropriate. In particular all "unknown" counts are zero which
means all transactions were processed.