Change processing
Add udf command, "update_distribution_file". It takes an input file and
checks each transaction updating those without a USD value. And then
writes the results to the provided output file.

Now pdf command only takes the input file and assumes the file is good
and then processes each transaction and tracks the "current" value of
each asset and then at the end outputs a report of the final end of year
value for each asset.