Need to work around mutating multiple entries in AssetRecMap
Mutating multiple entries of a collection by using references to them
violates the single mutator guarantee in Rust. Fix for now is to add
associated functions to AssetRecMap which gets a mutable reference,
mutates a field and then drops the reference. Thus only one reference
is held at a time.