Add get_asset_quantity_usd_value
The get_asset_quantity_usd_value takes into account the category and
operation when deciding what asset is the "asset" being more consistent.
In particular now when selling the quote_asset is the main asset instead
of the base_asset for both buying and selling.

This make the transaction_count field for each asset the same between what
is reported by ttffbudf and pttf. It isn't super significant as the
main thing is the quantity didn't change.

Also some tweaking of names.