Sort the ttr_vec data before outputing
This is what I'm doing in binance_us_processing so the same should be
done here. The reason was that in binance_us_processing the input files
can be in any order. Although actually processing multiple files isn't
typical it could happen and there is no real guarantee the input data
is in order so this is a good idea.