Fix bug where coin had no fiat price on TokenTax
I was using the start_period time but this doesn't work if the coin
didn't "exist" at that time. Instead I'll use first_dt, alternatively
we could use the last or sometime between the first and last.

In addition set cur_ttr and first_dt when we end the period and
if it's the first TypeTxs::Income transaction.

Also removed {consolidate_}push_to_consolidated_ttr_vec and just duplicate
the code since it makes it easier to see they are the same. Note we
couldn't use push_to_consolidated_ttr_vec inside the loop because it
causes E0502 compile error.