Changed TaxBitRec to use se_time_ms_to_utc_z_string for serialization.
This now meets TaxBit's specification. So now we'll output a trailing "Z"
instead of "+00:00" although in actuality TaxBit does support the
"+00:00" just fine. But there documentation doesn't say that so tech
support thinks the format it wrong.

We'll see if they complain about me using milli-seconds but they use that
format in their export format, so hopefully they'll accept it.