binance_withdraw: Change how withdraw address is passed
It can either be in the configuration.toml file as withdraw_addr or
as a command line option --withdraw-addr. Also changed is type in
WithdrawParams to Option<String> as it's quasi optional in that it doesn't
have to be a parameter but can be in the configuration file.

Also removed Configuration xyz: Option<String>, it isn't used.