refactor: Fix nightly clippy errors.
I was getting errors on the nightly build in github CI, so I ran
`cargo +nightly clippy` locally to see the errors see and fix the errors.

The two main types of errors was my technique of using a "let_return" style
in my code:

  fn blah() -> Z {
    ...

    let xxx = some expression

    xxx
  }

I do this because it makes it more obvious to me and allow me to add
logging of the result of the function return value at the end of the
function rather than multiple time if using an if or match statement.
To fix these I add #[allow(clippy::let_and_return)] at the expression:

  fn blah() -> Z {
    ...

    #[allow(clippy::let_and_return)]
    let xxx = some expression

    xxx
  }

The other errors were mostly requesting to derive Eq when deriving
PartialEq.