feat: Added SymbolFilters::TrailingDelta
This is a new field in response_body that wasn't parsed correct using
serde causing the error:

 wink@sellserver 22-11-30T22:06:05.128Z:~/binance-cli/data (main)
 $ cat binance-cli.cronjob_date_2022-11-01T00\:00\:01+00\:00.log
 binance-cli.work.sell-all.but-usd-bnb-eth.to-usd.buy-eth-with-usd.withdraw-eth.sh Starting at date=2022-11-01T00:01:37+00:00
 Error: Error("unknown variant `TRAILING_DELTA`, expected one of `PRICE_FILTER`, `PERCENT_PRICE`, `LOT_SIZE`, `MARKET_LOT_SIZE`, `MIN_NOTIONAL`, `ICEBERG_PARTS`, `MAX_NUM_ICEBERG_ORDERS`, `MAX_NUM_ORDERS`, `MAX_NUM_ALGO_ORDERS`, `MAX_POSITION`", line: 1, column: 1436)

The fix was modify binance_exchange_info.rs to have TrailingDeltaRec and
in SymbolFilters adding TrailingDelta which is associated the
TRAILING_DELTA and causes it to be renamed during serde processing.

I also temporarily added my custom-logger so I could quickly find the
error. This caused changes to Cargo.toml, rust-toolchain.toml and main.rs
with commented out changes so it's quicker next time I run into this.
I've commented them out as I didn't feel it a good idea to use "nightly"
toolchain for a "shipping" product :)